requirements.txt
-i https://pypi.org/simple
asgiref==3.8.1; python_version >= '3.8'
certifi==2025.1.31; python_version >= '3.6'
cffi==1.17.1; python_version >= '3.8'
charset-normalizer==3.4.1; python_version >= '3.7'
cryptography==44.0.2; python_version >= '3.7' and python_full_version not in '3.9.0, 3.9.1'
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
dj-database-url==3.0.0
django==4.2.23; python_version >= '3.8'
django-cors-headers==4.7.0; python_version >= '3.9'
django-filter==25.1; python_version >= '3.9'
djangorestframework==3.15.2; python_version >= '3.8'
djangorestframework-simplejwt==5.5.0; python_version >= '3.9'
djoser==2.3.1; python_version >= '3.8' and python_version < '4.0'
gunicorn==23.0.0; python_version >= '3.7'
idna==3.10; python_version >= '3.6'
oauthlib==3.2.2; python_version >= '3.6'
packaging==25.0; python_version >= '3.8'
pillow==11.2.1; python_version >= '3.9'
psycopg==3.2.5; python_version >= '3.8'
pycparser==2.22; python_version >= '3.8'
pyjwt==2.9.0; python_version >= '3.8'
python-dotenv==1.1.0; python_version >= '3.9'
python3-openid==3.2.0
requests==2.32.3; python_version >= '3.8'
requests-oauthlib==2.0.0; python_version >= '3.4'
social-auth-app-django==5.4.3; python_version >= '3.9'
social-auth-core==4.5.6; python_version >= '3.9'
sqlparse==0.5.3; python_version >= '3.8'
typing-extensions==4.14.0; python_version >= '3.9'
urllib3==2.3.0; python_version >= '3.9'
whitenoise==6.9.0; python_version >= '3.9'
